+++ src/main/java/org/team10424102/blackserver/models/Netbar.java
@@ -8,7 +8,7 @@ import javax.persistence.*;
 public class Netbar {
     @Id
     @GeneratedValue
-    private Integer id;
+    private Long id;
     private String name;
     @ManyToOne
     @JoinColumn(name = "logo_id")
@@ -17,11 +17,12 @@ public class Netbar {
 
     //<editor-fold desc="=== Getters & Setters ===">
 
-    public Integer getId() {
+
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
